# project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/bikes": {
            "get": {
                "tags": [
                    "Bikes"
                ],
                "summary": "List Bikes",
                "responses": {
                    "200": {
                        "description": "Success: List all Bikes",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Create Bike",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Bike",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/bikes/{id}": {
            "get": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Get Bike by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified bike by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Bike",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Update Bike",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified bike by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Bike updated",
                        "schema": {
                            "$ref": "#/definitions/Bike"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bikes"
                ],
                "summary": "Delete bike",
                "description": "Delete the specified bike by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bike id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Bike": {
            "required": [
                "make",
                "model",
                "year",
                "mods"
            ],
            "properties": {
                "make": {
                    "description": "Company name",
                    "type": "string",
                    "example": "Harley Davidson, Honda, Yamaha"
                },
                "model": {
                    "description": "Motorcycle model",
                    "type": "string",
                    "example": "Xl1200, Shadow ACE, V-Star"
                },
                "year": {
                    "description": "Fabrication year",
                    "type": "string",
                    "example": "2009, 2008, 2007"
                },
                "mods": {
                    "description": "Motorcycle description of modifications",
                    "type": "string",
                    "example": "New exhaust system"
                },
                "picture": {
                    "description": "Bike image URL",
                    "type": "string",
                    "example": "http://www.sample.com/my.bike.jpg"
                }
            }
        },
...
    }
}